finalizando

+++ Model/UsuarioService.py
@@ -12,16 +12,26 @@ def buscarUsuario(id: int):
 def salvar(user: UsuarioEntenty.Usuario):
     if user is None:
         return {'Usuario': 'Nao pode ser vazio'}
+
+    usuario = buscarUsuario(user.user_id)
+    if usuario != []:
+        return {'Usuario': 'Esse id ja foi cadastrado.'}
     return UsuarioRepository.save(user)
 
 
 def atualizar(user : UsuarioEntenty.Usuario):
     if user is None:
         return {'Usuario': 'Nao pode ser vazio'}
+    usuario = buscarUsuario(user.user_id)
+    if usuario == []:
+        return {'Usuario': 'Esse id nao existe.'}
     return UsuarioRepository.update(user)
 
 
 def deletar(id : int):
     if id is None:
         return {'Usuario': 'Nao pode ser vazio'}
+    usuario = buscarUsuario(id)
+    if usuario == []:
+        return {'Usuario': 'Esse id nao existe.'}
     return UsuarioRepository.delete(id)
